Solving merge conflicts
