+++ libs/core/schedulers/include/hpx/schedulers/local_priority_queue_scheduler.hpp
@@ -1263,7 +1263,7 @@ namespace hpx::threads::policies {
         // be terminated (i.e. no more work has to be done).
         bool wait_or_add_new(std::size_t num_thread, bool running,
             std::int64_t& idle_loop_count, bool enable_stealing,
-            std::size_t& added, thread_id_ref_type* = nullptr) override
+            std::size_t& added, thread_id_ref_type* = nullptr)
         {
             bool result = true;
